Update CI config, adds a parallel "lint" step, and mega-lint to fix rubocop issues.

+++ db/migrate/20230705195409_add_created_at_index_to_dependencies.rb
@@ -1,24 +1,17 @@
+# frozen_string_literal: true
+
 class AddCreatedAtIndexToDependencies < ActiveRecord::Migration[5.2]
   disable_ddl_transaction!
 
   def up
-    unless index_name_exists?(:dependencies, "index_dependencies_on_project_created_at_date")
-      add_index :dependencies, "project_id, (created_at::date)", name: "index_dependencies_on_project_created_at_date", algorithm: :concurrently
-    end
+    add_index :dependencies, "project_id, (created_at::date)", name: "index_dependencies_on_project_created_at_date", algorithm: :concurrently unless index_name_exists?(:dependencies, "index_dependencies_on_project_created_at_date")
 
-    unless index_name_exists?(:repository_dependencies, "index_repository_dependencies_on_project_created_at_date")
-      add_index :repository_dependencies, "project_id, (created_at::date)", name: "index_repository_dependencies_on_project_created_at_date", algorithm: :concurrently
-    end
-
+    add_index :repository_dependencies, "project_id, (created_at::date)", name: "index_repository_dependencies_on_project_created_at_date", algorithm: :concurrently unless index_name_exists?(:repository_dependencies, "index_repository_dependencies_on_project_created_at_date")
   end
 
   def down
-    if index_name_exists?(:dependencies, "index_dependencies_on_project_created_at_date")
-      remove_index :dependencies, name: "index_dependencies_on_project_created_at_date"
-    end
+    remove_index :dependencies, name: "index_dependencies_on_project_created_at_date" if index_name_exists?(:dependencies, "index_dependencies_on_project_created_at_date")
 
-    if index_name_exists?(:repository_dependencies, "index_repository_dependencies_on_project_created_at_date")
-      remove_index :repository_dependencies, name: "index_repository_dependencies_on_project_created_at_date"
-    end
+    remove_index :repository_dependencies, name: "index_repository_dependencies_on_project_created_at_date" if index_name_exists?(:repository_dependencies, "index_repository_dependencies_on_project_created_at_date")
   end
 end
